Тестирование функционала загрузки фотографии в каточку заметки.

+++ app/src/main/java/com/mu/fishermannotes/data/repository/NoteRepositoryImpl.kt
@@ -17,7 +17,7 @@ class NoteRepositoryImpl(
         return dao.getPhotos(noteId)
     }
 
-    override suspend fun update(note: NoteEntity): Long {
+    override suspend fun update(note: NoteEntity): Int {
         return dao.update(note)
     }
